Fix performance of list queries
Ordering items based on the highest timestamp found in the
changes table turned out to be slow. The initial
implementation looked at too many rows and produced 2 second
queries with 300.000 rows.

The initial plan to use CTE worked out for all queries expect
`GetAddedItems`. Explaining the queries showed that each using
a `WHERE recorded_at < ?` condition prompted SQLite to do a
`SEARCH` instruction. Only the single query not using a time
based condition used an expensive `SCAN` condition.

I was unable to write a query that would not show this behaviour.

To improve performance the timestamp of the last change is now
directly stored on the items. This allows all queries to be fast.