Fix migration that uses `PRAGMA foreign_keys`
This PRAGMA is a no-op inside transactions therefore the migration
does not work. To accommondate our migration logic we have to
commit/begin the wrapping transaction. A very dirty solution but
currently the only fast solution I can think of.

In the future a migration script should probably control the logic
around transactions. Atleast that is what other tools seem to be
doing.