Add webmanifest and homescreen icon
Adding Besserliste to the homescreen of ones mobilephone improves
the user experience. The icon and webmanifest ensure that the
mobile operating system treats Besserliste as an app as much as
possible.

While testing the process an error appeared. The `de_AT` collation
couldn't be loaded. I have currently no idea how that could happen
and was unable to reproduce it.